\name{MC.calibra-methods}
\docType{methods}
\alias{MC.calibra-methods}
\alias{MC.calibra,list,inputObs,numeric-method}
\alias{MC.calibra}
\title{Methods for Function \code{MC.calibra}}
\description{
Given the arguments of the method a calibration routine takes place. Method used only for internal purpose.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "list", obs = "inputObs", EmiStatR.cores = "numeric")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
