\name{stUPscales-package}
\alias{stUPscales-package}
\alias{stUPscales}
\docType{package}
\title{Spatio-Temporal Uncertainty Propagation Across Scales}
\description{The Spatio-Temporal Uncertainty Propagation Across Scales, stUPscales, 
    provides several R functions for 
    temporal aggregation of environmental variables used in Urban Drainage Models (UDMs), 
    as precipitation and pollutants.
    Also, it provides methods and functions for uncertainty propagation via Monte Carlo simulation. 
    This package, moreover, provides specific analysis functions for urban drainage 
    system simulation to evaluate water quantity and quality in combined sewer overflows (CSOs).     
}
\details{

The DESCRIPTION file:
\tabular{ll}{
Package: \tab stUPscales  \cr
Type:    \tab Package   \cr
Version: \tab 1.0.2       \cr
Date:    \tab 2017-10-04\cr
License: \tab GPL (>= 3)\cr
Depends: \tab R (>= 2.10), methods, mAr, data.table, lmom, xts \cr
}
}
\author{
    J.A. Torres-Matallana [aut, cre];
    U. Leopold [ctb];
    G.B.M. Heuvelink [ctb].

Maintainer: J.A. Torres-Matallana.
}

%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
