\name{Validation_Quantity-methods}
\docType{methods}
\alias{Validation_Quantity-methods}
\alias{Validation_Quantity,input,inputObs-method}
\alias{Validation_Quantity}
\title{Methods for Function \code{Validation_Quantity}}
\description{
Given the arguments of the method a validation routine takes place. Method used only for internal purpose.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "input", y = "inputObs")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}

